requirements.txt
Flask==0.10.1
Flask-DebugToolbar==0.9.0
Flask-SQLAlchemy==2.0
Jinja2==2.7.3
MarkupSafe==0.23
PyYAML==3.11
SQLAlchemy==0.9.7
Unidecode==0.04.16
Werkzeug==0.9.6
awesome-slugify==1.6
blinker==1.3
itsdangerous==0.24
regex==2014.09.18
